// src/models/Profile.ts
export interface Profile {
    user_id: number;
    firstname: string;
    lastname: string;
    dob: string;
    race: string;
    nationality: string;
    gender: string;
    status: string;
    last_login: string;
    type: string;
    address: string;
    lat: string;
    lng: string;
    title: string;
    email: string;
    password: string;
    phone: string;
    picture: string;
}